crear archivos rt en la carpeta routes
