+++ app/__init__.py
@@ -2,7 +2,8 @@ from flask import Flask
 from flask_migrate import Migrate
 from app.config import Config
 from app.db.sql import db, migrate
-from .utils.first_user import first_user  # 👈 import correcto
+from .utils.first_user import first_user
+from app.routes import register_routes  # 👈 Importa la función que registra todo
 
 def create_app():
     app = Flask(__name__)
@@ -11,13 +12,8 @@ def create_app():
     db.init_app(app)
     migrate.init_app(app, db)
 
-    # registrar blueprints y rutas cuando se tenga
-    # from app.routes import main_bp
-    # app.register_blueprint(main_bp)
-    
-    from app.routes import index_bp
-    app.register_blueprint(index_bp)
-
+    # Registrar TODAS las rutas
+    register_routes(app)
 
     with app.app_context():
         db.create_all()
