+++ app/templates/components/navbar.jinja2
@@ -6,34 +6,34 @@
 
         <div class="menu-icons">
             <div class="icon-item">
-                <a href="{{ url_for('IndexRoute.inicio') }}" style="text-decoration: none; color: inherit;">
+                <a href="{{ url_for('InicioRoute.inicio') }}" style="text-decoration: none; color: inherit;">
                     <span>🏠</span><br><span>Inicio</span>
                 </a>
             </div>
             <div class="icon-item">
-                <a href="{{ url_for('IndexRoute.empleos') }}" style="text-decoration: none; color: inherit;">
+                <a href="{{ url_for('EmpleosRoute.empleos') }}" style="text-decoration: none; color: inherit;">
                     <span>👥</span><br><span>Empleos</span>
                 </a>
             </div>
             <div class="icon-item">
-                <a href="{{ url_for('IndexRoute.mensajes') }}" style="text-decoration: none; color: inherit;">
+                <a href="{{ url_for('MensajesRoute.mensajes') }}" style="text-decoration: none; color: inherit;">
                     <span>💬</span><br><span>Mensajes</span>
                 </a>
             </div>
             <div class="icon-item">
-                <a href="{{ url_for('IndexRoute.notificaciones') }}" style="text-decoration: none; color: inherit;">
+                <a href="{{ url_for('NotificacionesRoute.notificaciones') }}" style="text-decoration: none; color: inherit;">
                     <span>🔔</span><br><span>Notificaciones</span>
                 </a>
             </div>
 
             <div class="icon-item">
-                <a href="{{ url_for('IndexRoute.planes') }}" style="text-decoration: none; color: inherit;">
+                <a href="{{ url_for('PlanesRoute.planes') }}" style="text-decoration: none; color: inherit;">
                     <span>💼</span><br><span>Planes</span>
                 </a>
             </div>
 
             <div class="icon-item">
-                <a href="{{ url_for('IndexRoute.logout') }}" style="text-decoration: none; color: inherit;">
+                <a href="{{ url_for('LoginRoute.logout') }}" style="text-decoration: none; color: inherit;">
                     <span>🚪</span><br><span>Salir</span>
                 </a>
             </div>
