+++ app/templates/empleos/empleos.jinja2
@@ -27,24 +27,6 @@
         </select>
     </div>
 </div>
-
-
-    <div class="content">
-        <h2 class="result-header">Empleos</h2>
-        <p class="result-sub">774 resultados para <strong>"desarrollador"</strong> en Chiapas</p>
-
-        <!-- Oferta destacada -->
-        <div class="job-card">
-            <div class="job-info">
-                <span class="job-label">Destacado</span>
-                <span class="job-title">Desarrollador Backend PHP</span>
-                <span class="job-meta">Nova Studio · Tuxtla Gutiérrez</span>
-            </div>
-            <button class="job-button">Ver detalles</button>
-        </div>
-
-    </div>
-
     {{ footer.footer() }}
 </body>
 </html>
