+++ app/templates/index.jinja2
@@ -11,7 +11,7 @@
     <header>
         <div class="logo"></div>
         <h1>WorkLink</h1>
-        <a href="{{ url_for('IndexRoute.login') }}" class="btn-login">Iniciar sesión</a>
+        <a href="{{ url_for('LoginRoute.login_form') }}" class="btn-login">Iniciar sesión</a>
     </header>
 
     <!-- Sección principal -->
