+++ app/templates/login.jinja2
@@ -39,7 +39,7 @@
                 <button type="submit" class="btn-entrar">Aceptar</button>
             </form>
 
-            <form method="get" action="{{ url_for('IndexRoute.registro') }}">
+            <form method="get" action="{{ url_for('RegistroRoute.registro') }}">
                 <button type="submit" class="btn-registrarse">Registrarse</button>
             </form>
