+++ app/templates/notificacion/notificaciones.jinja2
@@ -11,30 +11,6 @@
 
     {{ components.navbar("notificaciones", "notificaciones", current_user) }}
 
-    <div class="container">
-        <h2>Notificaciones</h2>
-
-        <div class="filtros">
-            <button class="filtro-btn">Todo</button>
-            <button class="filtro-btn">Vacantes</button>
-            <button class="filtro-btn">Estado</button>
-        </div>
-
-        <div class="notificacion">
-            <div class="icono">✅</div>
-            <button>Ver vacantes</button>
-        </div>
-
-        <div class="notificacion">
-            <div class="icono">➡️</div>
-            <button>Ver estado</button>
-        </div>
-
-        <div class="notificacion">
-            <div class="icono">⭐</div>
-            <button>Ver detalles</button>
-        </div>
-    </div>
 {{ footer.footer() }}
 </body>
 </html>
