+++ app/templates/planes/planes.jinja2
@@ -11,43 +11,6 @@
 
     {{ components.navbar("planes", "planes", current_user) }}
 
-    <div class="planes-container">
-        <h2 class="planes-title">Contamos con tres planes</h2>
-
-        <div class="plan-row">
-            <!-- Plan Gratuito -->
-            <div class="plan-card">
-                <div class="plan-title">Gratuito</div>
-                <div class="plan-icon">$</div>
-                <div class="plan-line"></div>
-                <button class="plan-button">continuar</button>
-            </div>
-
-            <!-- Plan Básico -->
-            <div class="plan-card">
-                <div class="plan-title">Basico</div>
-                <div class="plan-icon">$</div>
-                <div class="plan-line"></div>
-                <div class="plan-line"></div>
-                <div class="plan-line"></div>
-                <div class="plan-line"></div>
-                <button class="plan-button">Obtener</button>
-            </div>
-
-            <!-- Plan Premium -->
-            <div class="plan-card">
-                <div class="plan-title">Premium</div>
-                <div class="plan-icon">$</div>
-                <div class="plan-line"></div>
-                <div class="plan-line"></div>
-                <div class="plan-line"></div>
-                <div class="plan-line"></div>
-                <div class="plan-line"></div>
-                <div class="plan-line"></div>
-                <button class="plan-button">Obtener</button>
-            </div>
-        </div>
-    </div>
 
     {{ footer.footer() }}
 </body>
